refactor(contracts): split ipc-conversation.ts by domain
`ipc-conversation.ts` held ~15 unrelated domains in 1906 lines. The problem was
not size but the distance between an interface and the runtime guard that
validates it at the trust boundary: `AgentSummary` at 164, `isAgentSummary` at
1637, a gap of 1473 lines. `SidebarLayoutSnapshot` and its guard were 1413 apart,
`AttachmentSummary` 1130, `QueueSnapshot` 990.

That distance was not sloppiness. Lines 1471-1896 were one 425-line block serving
`isAgentEvent`, a 24-variant switch reaching into ~10 domains, so each guard sat
beside its caller rather than its type.

Replace it with fourteen domain modules whose imports flow strictly downward,
plus a package-internal `ipc-bounded-values.ts` for the `INPUT_LIMITS`-aware
primitives. Those do not go in `runtime-values.ts`, which is deliberately
limit-free and imported by renderer, backend and preload alike.
`ipc-agent-events.ts` is the only fan-in module and nothing imports it back, so
no cycle is possible.

`ipc-conversation.ts` was never in the package `exports` map, so all 355
importers go through the `@openbot/contracts/ipc` barrel and none of them change.
The public surface grows by exactly five guards that now have to cross a module
line; the six bounded-value helpers stay package-internal. Nothing else is added,
removed, or renamed.

`ipc.test.ts` imports the barrel and passes unedited, which is the verification a
pure move needs.
