+++ packages/contracts/src/ipc-agent-memories.ts
@@ -1,0 +1,52 @@
+import { INPUT_LIMITS } from "./input-limits";
+import { isDynamicRecord, isOneOf, isString } from "./runtime-values";
+
+export type AgentMemoryOrigin = "automatic" | "manual";
+
+export interface AgentMemory {
+  id: string;
+  agentId: string;
+  text: string;
+  origin: AgentMemoryOrigin;
+  sourceTurnId: string | null;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export function isAgentMemory(value: unknown): value is AgentMemory {
+  return (
+    isDynamicRecord(value) &&
+    isString(value.id) &&
+    value.id.length > 0 &&
+    value.id.length <= INPUT_LIMITS.identifier &&
+    isString(value.agentId) &&
+    value.agentId.length > 0 &&
+    value.agentId.length <= INPUT_LIMITS.identifier &&
+    isString(value.text) &&
+    value.text.length > 0 &&
+    value.text.length <= INPUT_LIMITS.agentMemoryText &&
+    isOneOf(["automatic", "manual"] as const, value.origin) &&
+    (value.sourceTurnId === null ||
+      (isString(value.sourceTurnId) &&
+        value.sourceTurnId.length > 0 &&
+        value.sourceTurnId.length <= INPUT_LIMITS.identifier)) &&
+    isString(value.createdAt) &&
+    isString(value.updatedAt)
+  );
+}
+
+export interface CreateAgentMemoryInput {
+  agentId: string;
+  text: string;
+}
+
+export interface UpdateAgentMemoryInput {
+  agentId: string;
+  memoryId: string;
+  text: string;
+}
+
+export interface DeleteAgentMemoryInput {
+  agentId: string;
+  memoryId: string;
+}
