+++ packages/contracts/src/ipc-approvals.ts
@@ -1,0 +1,89 @@
+import { INPUT_LIMITS } from "./input-limits";
+import { isBoundedString, isIdentifier, isNullableBoundedString, isRequestId } from "./ipc-bounded-values";
+import { isBoolean, isDynamicRecord, isOneOf } from "./runtime-values";
+
+export type AgentApprovalKind = "command" | "file-change" | "permissions";
+
+export interface AgentApprovalPermissions {
+  fileSystem: {
+    read: string[];
+    write: string[];
+  };
+  network: boolean;
+}
+
+function isAgentApprovalPermissions(value: unknown): value is AgentApprovalPermissions {
+  return (
+    isDynamicRecord(value) &&
+    isDynamicRecord(value.fileSystem) &&
+    isPathList(value.fileSystem.read) &&
+    isPathList(value.fileSystem.write) &&
+    isBoolean(value.network)
+  );
+}
+
+function isPathList(value: unknown): value is string[] {
+  return (
+    Array.isArray(value) &&
+    value.length <= INPUT_LIMITS.agents &&
+    value.every((path) => isBoundedString(path, INPUT_LIMITS.path))
+  );
+}
+
+export interface AgentApproval {
+  requestId: string | number;
+  agentId: string;
+  threadId: string;
+  turnId: string;
+  kind: AgentApprovalKind;
+  command: string | null;
+  cwd: string | null;
+  reason: string | null;
+  grantRoot: string | null;
+  permissions: AgentApprovalPermissions | null;
+}
+
+export function isAgentApproval(value: unknown): value is AgentApproval {
+  if (!isDynamicRecord(value)) return false;
+  return (
+    isRequestId(value.requestId) &&
+    isIdentifier(value.agentId) &&
+    isIdentifier(value.threadId) &&
+    isIdentifier(value.turnId) &&
+    isOneOf(["command", "file-change", "permissions"] as const, value.kind) &&
+    isNullableBoundedString(value.command, INPUT_LIMITS.messageText) &&
+    isNullableBoundedString(value.cwd, INPUT_LIMITS.path) &&
+    isNullableBoundedString(value.reason, INPUT_LIMITS.messageText) &&
+    isNullableBoundedString(value.grantRoot, INPUT_LIMITS.path) &&
+    (value.permissions === null || isAgentApprovalPermissions(value.permissions))
+  );
+}
+
+export interface RespondToApprovalInput {
+  requestId: string | number;
+  decision: "accept" | "decline";
+}
+
+export interface BrowserTakeoverRequest {
+  requestId: string | number;
+  agentId: string;
+  threadId: string;
+  turnId: string;
+  tabId: string;
+}
+
+export function isBrowserTakeoverRequest(value: unknown): value is BrowserTakeoverRequest {
+  return (
+    isDynamicRecord(value) &&
+    isRequestId(value.requestId) &&
+    isIdentifier(value.agentId) &&
+    isIdentifier(value.threadId) &&
+    isIdentifier(value.turnId) &&
+    isIdentifier(value.tabId)
+  );
+}
+
+export interface RespondToBrowserTakeoverInput {
+  requestId: string | number;
+  decision: "complete" | "cancel";
+}
