+++ packages/contracts/src/ipc-attachments.ts
@@ -1,0 +1,76 @@
+import { INPUT_LIMITS } from "./input-limits";
+import { isBoundedString, isIdentifier } from "./ipc-bounded-values";
+import { isDynamicRecord, isNumber, isOneOf } from "./runtime-values";
+
+export type AttachmentKind = "image" | "file";
+export type AttachmentPreviewKind = "image" | "pdf" | "text" | "none";
+
+export interface AttachmentSummary {
+  id: string;
+  name: string;
+  size: number;
+  kind: AttachmentKind;
+  mimeType: string;
+  previewKind: AttachmentPreviewKind;
+  previewUrl: string | null;
+}
+
+export function isAttachmentSummary(value: unknown): value is AttachmentSummary {
+  return (
+    isDynamicRecord(value) &&
+    isIdentifier(value.id) &&
+    isBoundedString(value.name, INPUT_LIMITS.attachmentName) &&
+    isNumber(value.size) &&
+    value.size >= 0 &&
+    isOneOf(["image", "file"] as const, value.kind) &&
+    isBoundedString(value.mimeType, INPUT_LIMITS.mimeType) &&
+    isOneOf(["image", "pdf", "text", "none"] as const, value.previewKind) &&
+    (value.previewUrl === null || isBoundedString(value.previewUrl, INPUT_LIMITS.avatarUrl))
+  );
+}
+
+export type DraftAttachment = AttachmentSummary;
+
+export interface ChooseAttachmentsInput {
+  filter: "all" | "images";
+}
+
+export interface AttachmentDataInput {
+  name: string;
+  mimeType: string;
+  bytes: Uint8Array;
+}
+
+export interface ImportAttachmentsInput {
+  paths: string[];
+  data: AttachmentDataInput[];
+}
+
+export type AttachmentImportEvent =
+  | { type: "started"; requestId: string; serverId: string }
+  | { type: "completed"; requestId: string; serverId: string; attachments: DraftAttachment[] }
+  | { type: "error"; requestId: string; serverId: string; message: string };
+
+export interface OpenAttachmentInput {
+  attachmentId: string;
+  action: "open" | "reveal" | "download";
+}
+
+export interface OpenSharedFileInput {
+  path: string;
+}
+
+export interface OpenWorkspaceFileInput {
+  agentId: string;
+  path: string;
+}
+
+export type FilePreviewKind = "markdown" | "text" | "image" | "pdf" | "none";
+
+export interface FilePreview {
+  name: string;
+  size: number;
+  mimeType: string;
+  previewKind: FilePreviewKind;
+  bytes: Uint8Array | null;
+}
