+++ packages/contracts/src/ipc-routines.ts
@@ -1,0 +1,215 @@
+import { INPUT_LIMITS } from "./input-limits";
+import { integerInRange } from "./ipc-bounded-values";
+import { isBoolean, isDynamicRecord, isOneOf, isString } from "./runtime-values";
+
+export type RoutineIntervalUnit = "minutes" | "hours" | "days";
+export type RoutineDaySelection =
+  | { kind: "every-day" }
+  | { kind: "days-of-week"; days: number[] }
+  | { kind: "days-of-month"; days: number[] };
+export type RoutineTimeSelection =
+  | { kind: "at-time"; time: string }
+  | { kind: "every"; amount: number; unit: Exclude<RoutineIntervalUnit, "days"> };
+
+export type RoutineSchedule =
+  | { kind: "hourly"; minute: number }
+  | { kind: "daily"; time: string }
+  | { kind: "weekdays"; time: string }
+  | { kind: "weekly"; weekday: number; time: string }
+  | { kind: "monthly"; day: number; time: string }
+  | { kind: "interval"; amount: number; unit: RoutineIntervalUnit; anchorAt: string }
+  | {
+      kind: "advanced";
+      months: number[];
+      days: RoutineDaySelection;
+      time: RoutineTimeSelection;
+    }
+  | { kind: "custom"; expression: string };
+
+export function isRoutineSchedule(value: unknown): value is RoutineSchedule {
+  if (!isDynamicRecord(value) || !isString(value.kind)) return false;
+  switch (value.kind) {
+    case "hourly":
+      return integerInRange(value.minute, 0, 59);
+    case "daily":
+    case "weekdays":
+      return isRoutineTime(value.time);
+    case "weekly":
+      return integerInRange(value.weekday, 0, 6) && isRoutineTime(value.time);
+    case "monthly":
+      return integerInRange(value.day, 1, 31) && isRoutineTime(value.time);
+    case "interval":
+      return (
+        integerInRange(value.amount, 1, 100_000) &&
+        isOneOf(["minutes", "hours", "days"] as const, value.unit) &&
+        isString(value.anchorAt) &&
+        !Number.isNaN(Date.parse(value.anchorAt))
+      );
+    case "advanced":
+      return (
+        Array.isArray(value.months) &&
+        value.months.length > 0 &&
+        value.months.every((month) => integerInRange(month, 1, 12)) &&
+        isRoutineDaySelection(value.days) &&
+        isRoutineTimeSelection(value.time)
+      );
+    case "custom":
+      return (
+        isString(value.expression) && value.expression.length > 0 && value.expression.length <= INPUT_LIMITS.routineCron
+      );
+    default:
+      return false;
+  }
+}
+
+function isRoutineTime(value: unknown): value is string {
+  return isString(value) && /^([01]\d|2[0-3]):[0-5]\d$/.test(value);
+}
+
+function isRoutineDaySelection(value: unknown): value is RoutineDaySelection {
+  if (!isDynamicRecord(value) || !isString(value.kind)) return false;
+  if (value.kind === "every-day") return true;
+  if (!Array.isArray(value.days) || value.days.length === 0) return false;
+  if (value.kind === "days-of-week") return value.days.every((day) => integerInRange(day, 0, 6));
+  if (value.kind === "days-of-month") return value.days.every((day) => integerInRange(day, 1, 31));
+  return false;
+}
+
+function isRoutineTimeSelection(value: unknown): value is RoutineTimeSelection {
+  if (!isDynamicRecord(value) || !isString(value.kind)) return false;
+  if (value.kind === "at-time") return isRoutineTime(value.time);
+  return (
+    value.kind === "every" &&
+    integerInRange(value.amount, 1, 100_000) &&
+    isOneOf(["minutes", "hours"] as const, value.unit)
+  );
+}
+
+export interface RoutineTrigger {
+  id: string;
+  routineId: string;
+  schedule: RoutineSchedule;
+  nextRunAt: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+function isRoutineTrigger(value: unknown): value is RoutineTrigger {
+  return (
+    isDynamicRecord(value) &&
+    isString(value.id) &&
+    isString(value.routineId) &&
+    isRoutineSchedule(value.schedule) &&
+    isString(value.nextRunAt) &&
+    isString(value.createdAt) &&
+    isString(value.updatedAt)
+  );
+}
+
+export interface Routine {
+  id: string;
+  agentId: string;
+  name: string;
+  instruction: string;
+  active: boolean;
+  timezone: string;
+  trigger: RoutineTrigger;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export function isRoutine(value: unknown): value is Routine {
+  return (
+    isDynamicRecord(value) &&
+    isString(value.id) &&
+    isString(value.agentId) &&
+    isString(value.name) &&
+    isString(value.instruction) &&
+    isBoolean(value.active) &&
+    isString(value.timezone) &&
+    isRoutineTrigger(value.trigger) &&
+    isString(value.createdAt) &&
+    isString(value.updatedAt)
+  );
+}
+
+export type RoutineRunStatus =
+  | "queued"
+  | "running"
+  | "needs-attention"
+  | "succeeded"
+  | "failed"
+  | "interrupted"
+  | "cancelled";
+
+export interface RoutineRun {
+  id: string;
+  routineId: string;
+  agentId: string;
+  triggerId: string | null;
+  kind: "scheduled" | "manual";
+  scheduledFor: string;
+  routineName: string;
+  instruction: string;
+  deliveryId: string | null;
+  status: RoutineRunStatus;
+  error: string | null;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export function isRoutineRun(value: unknown): value is RoutineRun {
+  return (
+    isDynamicRecord(value) &&
+    isString(value.id) &&
+    isString(value.routineId) &&
+    isString(value.agentId) &&
+    (value.triggerId === null || isString(value.triggerId)) &&
+    isOneOf(["scheduled", "manual"] as const, value.kind) &&
+    isString(value.scheduledFor) &&
+    isString(value.routineName) &&
+    isString(value.instruction) &&
+    (value.deliveryId === null || isString(value.deliveryId)) &&
+    isOneOf(
+      ["queued", "running", "needs-attention", "succeeded", "failed", "interrupted", "cancelled"] as const,
+      value.status,
+    ) &&
+    (value.error === null || isString(value.error)) &&
+    isString(value.createdAt) &&
+    isString(value.updatedAt)
+  );
+}
+
+export interface CreateRoutineInput {
+  agentId: string;
+  name: string;
+  instruction: string;
+  active: boolean;
+  timezone: string;
+  schedule: RoutineSchedule;
+}
+
+export interface UpdateRoutineInput {
+  agentId: string;
+  routineId: string;
+  name?: string;
+  instruction?: string;
+  active?: boolean;
+  schedule?: RoutineSchedule;
+}
+
+export interface DeleteRoutineInput {
+  agentId: string;
+  routineId: string;
+}
+
+export interface TestRoutineInput {
+  agentId: string;
+  routineId: string;
+}
+
+export interface ListRoutineRunsInput {
+  agentId: string;
+  routineId: string;
+  limit?: number;
+}
