+++ packages/contracts/src/ipc-sidebar-layout.ts
@@ -1,0 +1,46 @@
+import { isDynamicRecord, isNumber, isString } from "./runtime-values";
+
+export const SIDEBAR_PEOPLE_SECTION_ID = "people";
+export const SIDEBAR_UNASSIGNED_SECTION_ID = "unassigned";
+
+export interface SidebarSection {
+  id: string;
+  name: string;
+}
+
+export interface SidebarLayoutSnapshot {
+  revision: number;
+  sections: SidebarSection[];
+  order: string[];
+  agentAssignments: Record<string, string>;
+  agentOrder: string[];
+}
+
+export function isSidebarLayoutSnapshot(value: unknown): value is SidebarLayoutSnapshot {
+  if (!isDynamicRecord(value) || !isNumber(value.revision) || !Number.isInteger(value.revision) || value.revision < 0) {
+    return false;
+  }
+  if (
+    !Array.isArray(value.sections) ||
+    !Array.isArray(value.order) ||
+    !isDynamicRecord(value.agentAssignments) ||
+    !Array.isArray(value.agentOrder)
+  ) {
+    return false;
+  }
+  return (
+    value.sections.every((section) => isDynamicRecord(section) && isString(section.id) && isString(section.name)) &&
+    value.order.every(isString) &&
+    Object.values(value.agentAssignments).every(isString) &&
+    value.agentOrder.every(isString) &&
+    new Set(value.agentOrder).size === value.agentOrder.length
+  );
+}
+
+export type SidebarLayoutAction =
+  | { type: "create"; name: string; agentId?: string }
+  | { type: "rename"; sectionId: string; name: string }
+  | { type: "delete"; sectionId: string }
+  | { type: "move"; sectionId: string; direction: "up" | "down"; steps?: number }
+  | { type: "assign"; agentId: string; sectionId: string | null }
+  | { type: "move-agent"; agentId: string; sectionId: string | null; beforeAgentId: string | null };
