+++ packages/contracts/src/ipc.ts
@@ -1,11 +1,24 @@
+export * from "./ipc-agent-events";
+export * from "./ipc-agent-identity";
+export * from "./ipc-agent-memories";
+export * from "./ipc-agent-runtime";
+export * from "./ipc-agent-status";
+export * from "./ipc-agents";
 export * from "./ipc-app-auth";
+export * from "./ipc-approvals";
+export * from "./ipc-attachments";
 export * from "./ipc-browser";
 export * from "./ipc-channels";
-export * from "./ipc-conversation";
+export * from "./ipc-conversation-events";
+export * from "./ipc-conversation-messages";
+export * from "./ipc-conversations";
 export * from "./ipc-desktop-apis";
 export * from "./ipc-dynamic-island";
 export * from "./ipc-hosted-sites";
 export * from "./ipc-marketplace-agents";
+export * from "./ipc-queue";
+export * from "./ipc-routines";
+export * from "./ipc-sidebar-layout";
 export * from "./ipc-skills";
 export * from "./ipc-team-host";
 export * from "./ipc-voice";
